finished the login and sign up features of okta
