+++ client/package.json
@@ -5,6 +5,8 @@
   "private": true,
   "proxy": "http://localhost:3001/",
   "dependencies": {
+    "@okta/okta-react": "^1.2.0",
+    "@okta/okta-signin-widget": "^2.19.0",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
     "react-scripts": "^2.1.1"
