+++ client/src/components/layout/Navbar.js
@@ -6,7 +6,7 @@ class Navbar extends Component {
         <nav className="navbar navbar-expand-sm navbar-dark bg-dark mb-4">
         <div className="container">
           <Link className="navbar-brand" to="/">
-            Acme Staff Portal
+            BoozIt Members Portal
           </Link>
           <button
             className="navbar-toggler"
@@ -24,8 +24,8 @@ class Navbar extends Component {
                 </Link>
               </li>
               <li className="nav-item">
-                <Link className="nav-link" to="/staff">
-                  Staff
+                <Link className="nav-link" to="/Members">
+                  Members
                 </Link>
               </li>
             </ul>
