+++ client/src/components/models/beer.js
