--- src/App.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-
-import './App.css';
-
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <h1>Tell us About Your Most Recent Slug of Hooch</h1>
-      </div>
-    );
-  }
-}
-
-export default App;
